requirements.txt
flask
ultralytics
opencv-python
torch
torchvision
Pillow
numpy
gunicorn
gdown

